UI, Tables: move ViewControls into traits
