+++ components/ILIAS/UI/src/Implementation/Component/Table/Data.php
@@ -29,25 +29,23 @@ use ILIAS\UI\Component\Signal;
 use ILIAS\UI\Implementation\Component\JavaScriptBindable;
 use ILIAS\UI\Component\JavaScriptBindable as JSBindable;
 use ILIAS\Data\Factory as DataFactory;
-use ILIAS\Data\Order;
-use ILIAS\Data\Range;
+
+
 use ILIAS\UI\Component\Input\ViewControl;
 use ILIAS\UI\Component\Input\Container\ViewControl as ViewControlContainer;
-use ILIAS\UI\Implementation\Component\Input\ViewControl\Pagination;
 
 class Data extends AbstractTable implements T\Data
 {
     use JavaScriptBindable;
+    use TableViewControlFieldSelection;
+    use TableViewControlPagination;
+    use TableViewControlOrdering;
 
+    public const STORAGE_ID_PREFIX = self::class . '_';
     public const VIEWCONTROL_KEY_PAGINATION = 'range';
     public const VIEWCONTROL_KEY_ORDERING = 'order';
     public const VIEWCONTROL_KEY_FIELDSELECTION = 'selected_optional';
-    public const STORAGE_ID_PREFIX = self::class . '_';
 
-    protected int $number_of_rows = 800;
-
-    protected ?Range $range = null;
-    protected ?Order $order = null;
     protected ?array $filter = null;
     protected ?array $additional_parameters = null;
 
@@ -73,62 +71,14 @@ class Data extends AbstractTable implements T\Data
             $title,
             $columns
         );
-        $this->order = $this->data_factory->order($this->initialOrder(), Order::ASC);
-        $this->range = $data_factory->range(0, $this->number_of_rows);
-    }
-
-    private function initialOrder(): string
-    {
-        $visible_cols = $this->getVisibleColumns();
-        $sortable_visible_cols = array_filter(
-            $visible_cols,
-            static fn($c): bool => $c->isSortable()
-        );
-        if ($sortable_visible_cols === []) {
-            return array_key_first($visible_cols);
-        }
-        return array_key_first($sortable_visible_cols);
+        $this->initViewControlFieldSelection($columns);
+        $this->initViewControlOrdering();
+        $this->initViewControlpagination();
     }
 
     public function getDataRetrieval(): T\DataRetrieval
     {
         return $this->data_retrieval;
-    }
-
-    public function withNumberOfRows(int $number_of_rows): self
-    {
-        $clone = clone $this;
-        $clone->number_of_rows = $number_of_rows;
-        return $clone;
-    }
-
-    public function getNumberOfRows(): int
-    {
-        return $this->number_of_rows;
-    }
-
-    public function withOrder(?Order $order): self
-    {
-        $clone = clone $this;
-        $clone->order = $order;
-        return $clone;
-    }
-
-    public function getOrder(): Order
-    {
-        return $this->order ?? $this->data_factory->order($this->initialOrder(), Order::ASC);
-    }
-
-    public function withRange(?Range $range): self
-    {
-        $clone = clone $this;
-        $clone->range = $range;
-        return $clone;
-    }
-
-    public function getRange(): Range
-    {
-        return $this->range ?? $this->data_factory->range(0, $this->number_of_rows);
     }
 
     public function withFilter(?array $filter): self
@@ -162,7 +112,6 @@ class Data extends AbstractTable implements T\Data
             ->withSingleActions($this->getSingleActions())
             ->withVisibleColumns($this->getVisibleColumns());
     }
-
 
     /**
      * @return array<self, ViewControlContainer\ViewControl>
@@ -200,45 +149,4 @@ class Data extends AbstractTable implements T\Data
         $view_controls = array_filter($view_controls);
         return $this->view_control_container_factory->standard($view_controls);
     }
-
-    protected function getViewControlPagination(?int $total_count = null): ?ViewControl\Pagination
-    {
-        $smallest_option = current(Pagination::DEFAULT_LIMITS);
-        if (is_null($total_count) || $total_count >= $smallest_option) {
-            $range = $this->getRange();
-            return
-                $this->view_control_factory->pagination()
-                    ->withTotalCount($total_count)
-                    ->withValue([
-                        Pagination::FNAME_OFFSET => $range->getStart(),
-                        Pagination::FNAME_LIMIT => $range->getLength()
-                    ]);
-        }
-        return null;
-    }
-
-    protected function getViewControlOrdering(): ?ViewControl\Sortation
-    {
-        $sortable_visible_cols = array_filter(
-            $this->getVisibleColumns(),
-            static fn($c): bool => $c->isSortable()
-        );
-
-        if ($sortable_visible_cols === []) {
-            return null;
-        }
-
-        $sort_options = [];
-        foreach ($sortable_visible_cols as $col_id => $col) {
-
-            $order_asc = $this->data_factory->order($col_id, Order::ASC);
-            $order_desc = $this->data_factory->order($col_id, Order::DESC);
-
-            $labels = $col->getOrderingLabels();
-            $sort_options[$labels[0]] = $order_asc;
-            $sort_options[$labels[1]] = $order_desc;
-        }
-        return $this->view_control_factory->sortation($sort_options);
-    }
-
 }
